Added editing of UserOccasions. Link to User Occasion edit from SingleDayDispaly. Fixed bug in Month.js for MonthView when EbntryList is empty.

+++ App/Components/OccasionsScreen.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { ScrollView, Text, View, TouchableHighlight } from 'react-native';
+import { ScrollView, Text, View, TouchableHighlight, Alert } from 'react-native';
 import { Icon } from 'react-native-elements';
 import SideMenu from './SideMenu';
 import CustomList from './CustomList';
@@ -26,28 +26,42 @@ export default class OccasionsScreen extends Component {
         this.deleteOccasion = this.deleteOccasion.bind(this);
     }
     deleteOccasion(occasion) {
-        DataUtils.DeleteUserOccasion(occasion).then(() => {
-            const appData = this.state.appData;
-            let occasionList = appData.UserOccasions,
-                index = occasionList.indexOf(occasion);
-            if (index > -1) {
-                occasionList.splice(index, 1);
-                appData.UserOccasions = occasionList;
-                if (this.onUpdate) {
-                    this.onUpdate(appData);
-                }
-                this.setState({
-                    appData: appData,
-                    occasionList: occasionList
-                });
-                popUpMessage(`The ocassion ${occasion.title} has been successfully removed.`,
-                    'Remove occasion');
-            }
-        }
-        ).catch(err => {
-            warn('Error trying to delete an occasion from the database.');
-            error(err);
-        });
+        Alert.alert(
+            'Confirm Event Removal', 'Are you sue that you want to remove this Event?', [
+                //Button 1
+                {
+                    text: 'Cancel',
+                    onPress: () => { return; },
+                    style: 'cancel'
+                },
+                //Button 2
+                {
+                    text: 'OK', onPress: () => {
+                        DataUtils.DeleteUserOccasion(occasion).then(() => {
+                            const appData = this.state.appData;
+                            let occasionList = appData.UserOccasions,
+                                index = occasionList.indexOf(occasion);
+                            if (index > -1) {
+                                occasionList.splice(index, 1);
+                                appData.UserOccasions = occasionList;
+                                if (this.onUpdate) {
+                                    this.onUpdate(appData);
+                                }
+                                this.setState({
+                                    appData: appData,
+                                    occasionList: occasionList
+                                });
+                                popUpMessage(`The Event "${occasion.title}" has been successfully removed.`,
+                                    'Remove Event');
+                            }
+                        }
+                        ).catch(err => {
+                            warn('Error trying to delete an occasion from the database.');
+                            error(err);
+                            popUpMessage('We are sorry, Luach is unable to remove this Event.\nPlease contact lucah@compute.co.il.');
+                        });
+                    }
+                }]);
     }
     render() {
         return (
@@ -65,27 +79,49 @@ export default class OccasionsScreen extends Component {
                             emptyListText='There are no Events in the list'
                             secondSection={occasion => <View style={GeneralStyles.inItemButtonList}>
                                 <TouchableHighlight
-                                    underlayColor='#faa'
-                                    style={{ flex: 1 }}
-                                    onPress={() => this.deleteOccasion(occasion)}>
-                                    <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-                                        <Icon
-                                            name='delete-forever'
-                                            color='#faa'
-                                            size={25} />
-                                        <Text> Remove</Text>
-                                    </View>
-                                </TouchableHighlight>
-                                <TouchableHighlight
                                     underlayColor='#696'
                                     style={{ flex: 1 }}
-                                    onPress={() => this.navigate('Home', { currDate: occasion.jdate, appData: this.state.appData })}>
-                                    <View style={{ flexDirection: 'row', alignItems: 'center' }}>
+                                    onPress={() => this.navigate('Home', {
+                                        currDate: occasion.jdate,
+                                        appData: this.state.appData
+                                    })}>
+                                    <View style={{ alignItems: 'center' }}>
                                         <Icon
                                             name='event-note'
                                             color='#696'
-                                            size={25} />
-                                        <Text> Go to Date</Text>
+                                            size={18}
+                                            containerStyle={GeneralStyles.inItemLinkIcon} />
+                                        <Text style={GeneralStyles.inItemLinkText}> Go to Date</Text>
+                                    </View>
+                                </TouchableHighlight>
+                                <TouchableHighlight
+                                    underlayColor='#788778'
+                                    style={{ flex: 1 }}
+                                    onPress={() => this.navigate('NewOccasion', {
+                                        occasion,
+                                        appData: this.state.appData,
+                                        onUpdate: this.onUpdate
+                                    })}>
+                                    <View style={{ alignItems: 'center' }}>
+                                        <Icon
+                                            name='edit'
+                                            color='#99a999'
+                                            size={18}
+                                            containerStyle={GeneralStyles.inItemLinkIcon} />
+                                        <Text style={GeneralStyles.inItemLinkText}>Edit</Text>
+                                    </View>
+                                </TouchableHighlight>
+                                <TouchableHighlight
+                                    underlayColor='#faa'
+                                    style={{ flex: 1 }}
+                                    onPress={() => this.deleteOccasion(occasion)}>
+                                    <View style={{ alignItems: 'center' }}>
+                                        <Icon
+                                            name='delete-forever'
+                                            color='#faa'
+                                            size={18}
+                                            containerStyle={GeneralStyles.inItemLinkIcon} />
+                                        <Text style={GeneralStyles.inItemLinkText}> Remove</Text>
                                     </View>
                                 </TouchableHighlight>
                             </View>} />
